# Tekst-API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tekst-Dev",
    "description": "An online text research platform",
    "termsOfService": "None",
    "contact": {},
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1/api"
    }
  ],
  "paths": {
...
    "/locations": {
      "post": {
        "tags": [
          "locations"
        ],
        "summary": "Create location",
        "description": "Creates a new location. The position will be automatically set to the last position\nof the location's parent (or the first parent before that has children).",
        "operationId": "createLocation",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "locations"
        ],
        "summary": "Find locations",
        "operationId": "findLocations",
        "parameters": [
          {
            "name": "txt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "description": "ID of text to find locations for",
              "title": "Txt"
            },
            "description": "ID of text to find locations for"
          },
          {
            "name": "lvl",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Structure level to find locations for",
              "title": "Lvl"
            },
            "description": "Structure level to find locations for"
          },
          {
            "name": "pos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Position value of locations to find",
              "title": "Pos"
            },
            "description": "Position value of locations to find"
          },
          {
            "name": "parent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "description": "ID of parent location to find children of",
              "title": "Parent"
            },
            "description": "ID of parent location to find children of"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Return at most <limit> locations",
              "default": 8192,
              "title": "Limit"
            },
            "description": "Return at most <limit> locations"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationRead"
                  },
                  "title": "Response Find Locations Locations Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/locations/{id}": {
      "get": {
        "tags": [
          "locations"
        ],
        "summary": "Get location",
        "operationId": "getLocation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "locations"
        ],
        "summary": "Update location",
        "operationId": "updateLocation",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "locations"
        ],
        "summary": "Delete location",
        "description": "Deletes the specified location. Also deletes any associated contents, child locations and contents associated with child locations.",
        "operationId": "deleteLocation",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteLocationResult"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/texts": {
      "get": {
        "tags": [
          "texts"
        ],
        "summary": "Get all texts",
        "operationId": "getAllTexts",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 128,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TextRead"
                  },
                  "title": "Response Get All Texts Texts Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "texts"
        ],
        "summary": "Create text",
        "operationId": "createText",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/texts/{id}/template": {
      "get": {
        "tags": [
          "texts"
        ],
        "summary": "Download structure template",
        "description": "Download the structure template for a text to help compose a structure\ndefinition that can later be uploaded to the server",
        "operationId": "downloadStructureTemplate",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/texts/{id}/structure": {
      "post": {
        "tags": [
          "texts"
        ],
        "summary": "Import text structure",
        "description": "Upload the structure definition for a text to apply as a structure of locations",
        "operationId": "importTextStructure",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_text_structure_texts__id__structure_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_text_structure_texts__id__structure_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "JSON file containing the text's structure"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_text_structure_texts__id__structure_post"
      },
...
      "DeleteLocationResult": {
        "properties": {
          "contents": {
            "type": "integer",
            "title": "Contents"
          },
          "locations": {
            "type": "integer",
            "title": "Locations"
          }
        },
        "type": "object",
        "required": [
          "contents",
          "locations"
        ],
        "title": "DeleteLocationResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationCreate": {
        "properties": {
          "textId": {
            "type": "string",
            "title": "Textid",
            "description": "ID of the text this location belongs to",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "parentId": {
            "anyOf": [
              {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentid",
            "description": "ID of parent location"
          },
          "level": {
            "type": "integer",
            "exclusiveMaximum": 32.0,
            "minimum": 0.0,
            "title": "Level",
            "description": "Index of structure level this location is on"
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position",
            "description": "Position among all text locations on this level"
          },
          "label": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Label",
            "description": "Label for identifying this text location in level context"
          }
        },
        "type": "object",
        "required": [
          "textId",
          "level",
          "position",
          "label"
        ],
        "title": "LocationCreate"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tekstuserauth"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  },
...
}